Move colour decoding tests into own fixture
